requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
groq
pydantic==2.5.0
pymongo==4.7.2
python-jose[cryptography]==3.3.0
httpx==0.27.0
